Extracting the IServiceContext interface from the Scope class and basing the IContainer interface on the IServiceContext interface
